Audio Manager Advanced Setup
Setup the audio manager to play sounds in a more complex manner. Sounds can be set to overlap, overwrite, or not play. Sounds can be played in a time delay. Accepts a delegate to occur after the sound has played. Accepts a delegate to check if the sound should play.

Used the above to create a first-pass of what the newly added sounds would look like in game.